+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Landing bg image, bg color & font color are prismic bound now
 - Site description is dynamic
+- Site footer is prismic bound
 
 ## [2.1.2] - 2019-08-27
 ### Changed